Refactor members cards
