+++ hooks/useMembers.ts
@@ -3,11 +3,11 @@ import { useQuery } from "@tanstack/react-query";
 export const useMembers = () => {
   const { status, data, error } = useQuery({
     queryKey: ["members"],
-    queryFn: () => fetch("/api/members"),
+    queryFn: () => fetch("/api/members").then((res) => res.json()),
   });
 
   return {
-    members: data,
+    data,
     status,
     error,
   };
