+++ pages/api/members.ts
@@ -1,11 +1,20 @@
-import { NotionDB } from "notiondb";
-
 export default async function handler(req, res) {
-  const integrationKey = process.env.NOTION_INTEGRATION_KEY;
-  const notionDB = new NotionDB(integrationKey);
+  // TODO: This is mock data. Use notion db instead
+  const members = [
+    {
+      name: "David Podolskyi",
+      role: "President",
+      image: "https://www.tum-ai.com/assets/img/members/david_podolskyi.webp",
+      description:
+        "ML & Entrepreneurship enthusiast with focus on IoT and Recommender Systems",
+    },
+    {
+      name: "Hamze Al-Zamkan",
+      role: "President",
+      image: "https://www.tum-ai.com/assets/img/members/hamze_alzamkan.webp",
+      description: "lorem ipsum dolor sit amet",
+    },
+  ];
 
-  const databases = await notionDB.databases.getAll();
-  console.log(databases.map((d) => d.object));
-
-  res.status(200).json({ name: "John Doe" });
+  res.status(200).json(members);
 }
